pyproject.toml
[project]
name = "simpleagent-interview-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "simple-agents-py>=0.0.24",
  "simpleflow-sdk>=0.1.1",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
